Specify avrdude 6.3.0-arduino17 tool dependency for ATtinyCore
avrdude 6.3.0-arduino16 used by the current version of Arduino megaAVR Boards has a bug that affects the USBtinyISP programmer. Prior to this change, ATTinyCore would use the newest installed version of avrdude, so installing Arduino megaAVR Boards would break USBtinyISP for ATTinyCore. This bug in avrdude has been fixed in avrdude 6.3.0-arduino17. By specifying avrdude 6.3.0-arduino17 as a dependency, that version of avrdude will be installed during the Boards Manager installation and used when uploading to ATtinyCore. This does not fix the problem for existing Boards Manager installations and manual installations.